requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
django-taggit~=2.1.0
django-filter~=21.1
drf-spectacular~=0.21.2
drf-recaptcha~=2.0.7